+++ src/components/Table/index.js
@@ -154,7 +154,7 @@ type Props<T = Data> = {
   style?: Object,
   options?: TableOptions<T>,
   showLoader?: boolean,
-  showEmptyTable?: boolean,
+  renderEmptyTable?: boolean,
   tableButton?: {
     label: string,
     icon?: string | Node,
@@ -315,12 +315,12 @@ class Table<T> extends React.Component<Props<T>, State> {
       className,
       style,
       showLoader,
-      showEmptyTable,
+      renderEmptyTable,
       tableButton,
       options: { headerSection },
     } = this.props;
 
-    if (!data.length && !showEmptyTable) {
+    if (!data.length && !renderEmptyTable) {
       return null;
     }
 
@@ -400,7 +400,7 @@ Table.defaultProps = {
     headerSection: undefined,
   },
   showLoader: false,
-  showEmptyTable: false,
+  renderEmptyTable: false,
   tableButton: null,
 };