+++ src/index.d.ts
@@ -211,7 +211,7 @@ export interface TableProps<TableData extends TableRowData = TableRowData> {
   columns: Array<TableColumn<TableData>>;
   data: Array<TableData>;
   showLoader?: boolean;
-  showEmptyTable?: boolean;
+  renderEmptyTable?: boolean;
   className?: string;
   style?: CSSProperties;
   options?: Partial<TableOptions<TableData>>;